feat:CRUD de cadastro concluido, faltando apenas implementação da autenticação com Spring Security.

+++ src/main/java/com/codigoCerto/desafioBackEnd/service/impl/UserServiceImpl.java
@@ -7,6 +7,9 @@ import com.codigoCerto.desafioBackEnd.dto.request.UserSignUpRequest;
 import com.codigoCerto.desafioBackEnd.dto.response.UserEditProfileResponse;
 import com.codigoCerto.desafioBackEnd.dto.response.UserSignUpResponse;
 import com.codigoCerto.desafioBackEnd.entity.UserEntity;
+import com.codigoCerto.desafioBackEnd.exception.EmailAlreadyExists;
+import com.codigoCerto.desafioBackEnd.exception.IsIdStoredAtDataBase;
+import com.codigoCerto.desafioBackEnd.exception.IsPasswordEquals;
 import com.codigoCerto.desafioBackEnd.repository.impl.UserRepository;
 import com.codigoCerto.desafioBackEnd.service.IUserService;
 import lombok.RequiredArgsConstructor;
@@ -14,6 +17,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -23,6 +27,13 @@ public class UserServiceImpl implements IUserService {
 
     @Override
     public UserSignUpResponse createUser(UserSignUpRequest userSignUpRequest) {
+        UserEntity userEmailAlreadyRegistered = userRepository.findByEmail(userSignUpRequest.email());
+        if (userEmailAlreadyRegistered.getEmail() != null){
+            throw new EmailAlreadyExists("Email já cadastrado no sistema");
+        }
+        if (!Objects.equals(userSignUpRequest.password(), userSignUpRequest.confirmedPassword())){
+            throw new IsPasswordEquals("Senhas não iguais");
+        }
         UserEntity userTransformedToEntity = UserSignUpMapper.transformRequestToEntity(userSignUpRequest);
         UserEntity savedUserInRepo = this.userRepository.save(userTransformedToEntity);
         return UserSignUpMapper.transformEntityToResponse(savedUserInRepo);
@@ -37,11 +48,18 @@ public class UserServiceImpl implements IUserService {
     @Override
     public UserSignUpResponse getUserById(Long id) {
         UserEntity userReturnedFromRepo = userRepository.findById(id);
+        if (userReturnedFromRepo.getId() == null){
+            throw new IsIdStoredAtDataBase("Id não está armazenado no sistema");
+        }
         return UserSignUpMapper.transformEntityToResponse(userReturnedFromRepo);
     }
 
     @Override
     public UserEditProfileResponse updateUserById(Long id, UserEditProfileRequest userEditProfileRequest) {
+        UserEntity userReturnedFromRepo = userRepository.findById(id);
+        if (userReturnedFromRepo.getId() == null){
+            throw new IsIdStoredAtDataBase("Id não está armazenado no sistema");
+        }
         UserEntity userEntity = UserEditMapper.transformRequestToEntity(userEditProfileRequest);
         UserEntity userEntityUpdatedFromRepo = userRepository.updateById(id, userEntity);
         return UserEditMapper.transformEntityToResponse(userEntityUpdatedFromRepo);
@@ -49,6 +67,10 @@ public class UserServiceImpl implements IUserService {
 
     @Override
     public void deleteUserById(Long id) {
+        UserEntity userReturnedFromRepo = userRepository.findById(id);
+        if (userReturnedFromRepo.getId() == null){
+            throw new IsIdStoredAtDataBase("Id não está armazenado no sistema");
+        }
         this.userRepository.deleteById(id);
     }
 }
